pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "autotrader"
version = "0.1.0"
description = "AutoTrader - Cryptocurrency Trading Bot"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "ccxt>=4.0.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "aiohttp>=3.8.0",
    "prometheus-client>=0.17.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
microstructure = [
    "ccxt[async]>=4.0.0",  # WebSocket support via ccxt.pro
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "httpx>=0.24.0",  # For testing FastAPI endpoints
]
bouncehunter = [
    "yfinance>=0.2.40",
    "scikit-learn>=1.3.0",
    "lxml>=4.9.0",
    "requests>=2.31.0",
]

[project.scripts]
autotrader-backtest = "pipeline.cli_backtest:cli_main"
autotrader-backtest-harness = "backtest.harness:main"
autotrader-scan = "src.cli.main:cli_main"
autotrader-bouncehunter = "src.bouncehunter.cli:main"
autotrader-telegram-bot = "src.bouncehunter.telegram_cli:main"

[project.entry-points."autotrader.strategies"]
default = "src.core.pipeline:HiddenGemScanner"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "pipeline*", "backtest*"]
namespaces = false

[tool.pytest.ini_options]
addopts = "-q --maxfail=1"

[tool.coverage.run]
branch = true
source = ["src"]
omit = ["tests/*"]

[tool.coverage.report]
fail_under = 75
show_missing = true
skip_covered = true

[tool.coverage.html]
show_contexts = true
